kafka-connect/apps/transactions/requirements.txt
Faker
mysql-connector-python